requirements.txt
langchain_community
langchain
sentence-transformers==2.1.0
streamlit 
dotnev
google-generativeai
